requirements.txt
# Core
streamlit>=1.37.0
pandas>=2.2.2
gitpython>=3.1.43

# OpenAI
openai>=1.40.0


# Fix NumPy compatibility issues
numpy<2
numexpr>=2.9.0
bottleneck>=1.4.0

# Visualization (for commit timelines & analytics)
matplotlib>=3.9.0
plotly>=5.23.0
altair>=5.4.0

# Utility
python-dotenv>=1.0.1
